pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mint"
version = "0.1.0"
description = "Learning the language of protein-protein interactions"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Toby Hallet, Stefano Ubertti, Jakub Lala, Harsh Agarwal, Rishi Sunak"}
]
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "biopython>=1.79",
    "einops>=0.7.0",
    "fairscale>=0.4.6",
    "hydra-core>=1.3.2,<2.0",
    "omegaconf>=2.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "lightning>=2.2.0",
    "deepspeed>=0.10.0",
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.4",
    "sentencepiece>=0.2.0",
    "protobuf>=4.0.0",
    "requests>=2.28.0",
    "filelock>=3.12.2",
    "gdown>=4.7.3",
    "regex>=2024.4.16",
    "absl-py>=2.1.0",
    "openpyxl>=3.1.3",
    "rdkit>=2022.9.5",
    "biotite>=0.40.0",
    "fire>=0.7.0",
    "rcsbsearchapi<2.0",
    "wandb>=0.21.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "ipykernel>=6.25.0",
    "ipython>=8.15.0",
    "jupyter_client>=8.3.0",
    "jupyter_core>=5.3.0",
    "jupyterlab-widgets>=3.0.10",
]
bio = [
    "openmm>=8.0.0",
    "pdbfixer @ git+https://github.com/openmm/pdbfixer@1.9",
]

[project.urls]
Homepage = "https://github.com/softnanolab/mint"
Repository = "https://github.com/softnanolab/mint"
Issues = "https://github.com/softnanolab/mint/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 99
target-version = ["py39", "py310", "py311", "py312"]
exclude = '''
/(
    \.git
  | __pycache__
  | build
  | dist
  | experimental
  | third_party
)/
'''

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "extended: marks tests as extended (deselect with '-m \"not extended\"')"
]

[tool.coverage.run]
source = ["src/mint"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
